requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
--index-url https://in.thewardro.be/api/v4/projects/1129/packages/pypi/simple
--extra-index-url https://pypi.python.org/simple
--trusted-host in.thewardro.be
--trusted-host pypi.python.org

aiohttp==3.8.4
    # via gcloud-aio-auth
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
backoff==2.2.1
    # via gcloud-aio-auth
cffi==1.15.1
    # via cryptography
chardet==5.1.0
    # via gcloud-aio-auth
charset-normalizer==3.1.0
    # via aiohttp
cryptography==40.0.2
    # via gcloud-aio-auth
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gcloud-aio-auth==4.2.1
    # via gcloud-aio-pubsub
gcloud-aio-pubsub==5.4.0
    # via -r requirements.in
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
prometheus-client==0.16.0
    # via gcloud-aio-pubsub
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via gcloud-aio-auth
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
